requirements.txt
customtkinter>=5.2.0
Pygments>=2.15.0
tkcalendar>=1.6.1
argon2>=0.1.10
argon2-cffi>=20.1.0
Pillow==9.3.0